updated model for 2048 feature count
