+++ .gitignore
@@ -1,1 +1,3 @@
 .venv
+test.py
+uploads/downloaded_video.mp4
